+++ upfront-theme-exporter.php
@@ -3,7 +3,7 @@
 Plugin Name: Upfront Builder
 Plugin URI: http://premium.wpmudev.com/
 Description: The fastest, most visual, way to build WordPress themes. Now anyone can design, build, export, share, and sell WordPress themes.
-Version: 1.1.1
+Version: 1.1.8-BETA-2
 Author: WPMU DEV
 Text Domain: upfront_thx
 Author URI: http://premium.wpmudev.com
@@ -32,6 +32,9 @@ require_once dirname(__FILE__) . '/lib/util.php';
 require_once dirname(__FILE__) . '/lib/class_thx_l10n.php';
 
 define('THX_BASENAME', basename(dirname(__FILE__)));
+define('THX_PLUGIN_BASENAME', plugin_basename(__FILE__));
+
+register_deactivation_hook(__FILE__, array('UpfrontThemeExporter', 'deactivate'));
 
 /**
  * Main plugin class
@@ -44,6 +47,20 @@ class UpfrontThemeExporter {
 	 * Just basic, context-free bootstrap here.
 	 */
 	private function __construct() {}
+
+	/**
+	 * Plugin deactivation hook listener
+	 *
+	 * Cleans up stuff after itself.
+	 * Currently cleans up kicsktart activation notice with no
+	 * core present dismissal flag.
+	 *
+	 * @since 1.1.8-BETA-1
+	 */
+	public static function deactivate () {
+		if (!class_exists('Thx_Kickstart')) require_once dirname(__FILE__) . '/lib/class_thx_kickstart.php';
+		Thx_Kickstart::clean_up();
+	}
 
 	/**
 	 * Boot point.
@@ -111,10 +128,7 @@ class UpfrontThemeExporter {
 	 * No need to wait for the rest of Upfront, set our stuff up right now.
 	 */
 	private function _add_global_hooks () {
-		/*
-		// Not adding the toolbar item since the admin page move
 		add_action('upfront-admin_bar-process', array($this, 'add_toolbar_item'), 10, 2);
-		*/
 
 		add_filter('extra_theme_headers', array($this, 'process_extra_child_theme_headers'));
 
@@ -123,6 +137,62 @@ class UpfrontThemeExporter {
 			Thx_Admin::serve();
 		}
 		$this->_load_textdomain();
+
+		// Add shared Upfront/Exporter JS resources
+		add_action('upfront-core-inject_dependencies', array($this, 'add_shared_scripts'));
+		add_filter('upfront_data', array($this, 'add_shared_data'));
+
+		// Shared - context mode popup
+		add_action('wp_ajax_upfront_thx-mode_context-skip', array($this, 'json_context_mode_skip'));
+	}
+
+	/**
+	 * Handle data augmentation for shared resource
+	 *
+	 * @param array $data Data gathered this far
+	 *
+	 * @return array Augmented data
+	 */
+	public function add_shared_data ($data) {
+		$user_id = get_current_user_id();
+
+		$data['exporter_shared'] = array(
+			'context_mode' => (int)get_user_option('upfront-context_mode_skip', $user_id),
+		);
+
+		return $data;
+	}
+
+	/**
+	 * Stores user preference regarding context mode popup
+	 *
+	 * AJAX request handler
+	 */
+	public function json_context_mode_skip () {
+		$result = array('error' => 1);
+		if (!Upfront_Permissions::current(Upfront_Permissions::BOOT)) return wp_send_json($result);
+
+		$user_id = get_current_user_id();
+		if (empty($user_id) || !is_numeric($user_id)) return wp_send_json($result);
+
+		$skip = (int)get_user_option('upfront-context_mode_skip', $user_id);
+		if ($skip) return wp_send_json($result);
+
+		update_user_option($user_id, 'upfront-context_mode_skip', 1, true);
+		wp_send_json(array(
+			'error' => 0,
+		));
+	}
+
+	/**
+	 * This is where we inject shared scripts
+	 */
+	public function add_shared_scripts () {
+		if (!Upfront_Permissions::current(Upfront_Permissions::BOOT)) return false;
+
+		$deps = Upfront_CoreDependencies_Registry::get_instance();
+		$deps->add_script(plugins_url('app/shared.js', __FILE__));
+		$deps->add_style(plugins_url('styles/shared.css', __FILE__));
 	}
 
 	/**
@@ -173,46 +243,52 @@ class UpfrontThemeExporter {
 	/**
 	 * Adds the builder toolbar item
 	 *
-	 * Deprecated since v0.9
-	 *
 	 * @param object $toolbar Toolbar object
 	 * @param array $item Upfront item
 	 */
 	public function add_toolbar_item ($toolbar, $item) {
-		return false; // Deprecated since v0.9
-
 		if (!Upfront_Permissions::current(Upfront_Permissions::BOOT)) return false;
 		if (empty($item['meta'])) return false; // Only actual boot item has meta set
 
 		$child = upfront_thx_is_current_theme_upfront_child();
-		$create_title = __('Create New Theme', self::DOMAIN);
-		$main_title = (bool)$child
-			? __('Builder', self::DOMAIN)
-			: $create_title
-		;
-		$root_item_id = 'upfront-create-theme';
 
-		$toolbar->add_menu(array(
+		// Swap out root items
+		$root_item_id = $item['id'];
+		$new_item_root = 'upfront-editor_builder-hub';
+		$new_item = array_merge($item, array(
 			'id' => $root_item_id,
-			'title' => '<span style="top:2px" class="ab-icon dashicons-hammer"></span><span class="ab-label">' . $main_title . '</span>',
-			'href' => admin_url('admin.php?page=upfront-builder'),
-			'meta' => array( 'class' => 'upfront-create_theme' ),
+			'parent' => $new_item_root,
+		));
+		$item['id'] = $new_item_root;
+		unset($item['meta']);
+		unset($item['href']);
+
+		$toolbar->add_node($item);
+		$toolbar->add_node($new_item);
+
+		// Now, let's make builder menu item expandable
+		$toolbar->add_node(array(
+			'id' => 'upfront-builder-hub',
+			'parent' => $new_item_root,
+			'title' => __('Builder', self::DOMAIN),
 		));
 
 		if ((bool)$child) {
+			// Edit current
 			$toolbar->add_menu(array(
-				'parent' => $root_item_id,
+				'parent' => 'upfront-builder-hub',
 				'id' => 'upfront-builder-current_theme',
-				'title' => __('Edit current theme', self::DOMAIN),
+				'title' => __('Modify current theme', self::DOMAIN),
 				'href' => home_url('/' . UpfrontThemeExporter::get_root_slug() . '/' . $child),
 			));
-			$toolbar->add_menu(array(
-				'parent' => $root_item_id,
-				'id' => 'upfront-builder-create_theme',
-				'title' => $create_title,
-				'href' => admin_url('admin.php?page=upfront-builder'),
-			));
 		}
+		// Create new
+		$toolbar->add_menu(array(
+			'parent' => 'upfront-builder-hub',
+			'id' => 'upfront-builder-create_theme',
+			'title' => __('Create new theme', self::DOMAIN),
+			'href' => admin_url('admin.php?page=upfront-builder'),
+		));
 	}
 
 	/**