+++ CHANGELOG.md
@@ -2,6 +2,38 @@ Change Log
 ============
 
 
+1.1.7 - 2017-03-29
+-------------------------------------------------------------------------------
+- Add: compatibility with CoursePress core support
+- Add: admin toolbar combined menu item
+- Add: context mode popup
+
+1.1.6 - 2017-01-20
+-------------------------------------------------------------------------------
+- Fix: network activation issue
+
+1.1.5 - 2017-01-19
+-------------------------------------------------------------------------------
+- Add: supported video extensions to the image export list
+- Add: export to support save only on export
+- Fix: compatibility with Theme Tester
+- Fix: styles conflict with initial popup
+
+1.1.4 - 2016-11-23
+-------------------------------------------------------------------------------
+- Add: compatibility with MarketPress core support.
+
+1.1.3 - 2016-10-24
+-------------------------------------------------------------------------------
+- Add: compatibility with WooCommerce core support.
+
+1.1.2 - 2016-10-05
+-------------------------------------------------------------------------------
+- Add: compress data when exporting
+
+- Fix: disable esc key when on getting started screen
+
+
 1.1.1 - 2016-09-23
 -------------------------------------------------------------------------------
 - Fix: issue with global theme CSS selector prefixing